requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
supabase==1.0.4
openai==1.2.4
pytest==7.4.3
loguru==0.7.2 